Booking details updates

+++ src/main/java/com/smartstay/smartstay/responses/electricity/RoomElectricityList.java
@@ -6,6 +6,6 @@ public record RoomElectricityList(int ebId,
                                   Integer roomId,
                                   String startDate,
                                   String endDate,
-                                  Integer consumption,
-                                  Integer reading) {
+                                  Double consumption,
+                                  Double reading) {
 }
